# aspnetcore/src/IO.Swagger/Controllers/DefaultApi.cs
/*
 * Stories Collection
 *
 * Endpoints for updating Stories of Tales App
 *
 * OpenAPI spec version: 0.1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class DefaultApiController : ControllerBase
    { 
        /// <summary>
        /// 
        /// </summary>
        /// <remarks>Get all active stories of a set by ID</remarks>
        /// <param name="id">An ID to look up one or more values</param>
        /// <response code="200">OK</response>
        /// <response code="400">Bad Request</response>
        /// <response code="403">Access Denied</response>
        /// <response code="404">The resource is not found</response>
        [HttpGet]
        [Route("/api/Story/GetAllActiveStoriesOfSet/{id}")]
        [ValidateModelState]
        [SwaggerOperation("ApiStoryGetAllActiveStoriesOfSetIdGet")]
        [SwaggerResponse(statusCode: 200, type: typeof(ArrayOfStories), description: "OK")]
        public virtual IActionResult ApiStoryGetAllActiveStoriesOfSetIdGet([FromRoute][Required]string id)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(ArrayOfStories));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403);

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "[ {\n  \"StoryId\" : \"f89da017-05bb-48b0-c02f-08d80a57a67b\",\n  \"StorySetId\" : \"e520e6df-4ac0-49c8-7dcd-08d80a569eba\",\n  \"StoryName\" : \"Story about childhood\",\n  \"StoryCoverImage\" : \"/kids.png\",\n  \"IsActive\" : true\n}, {\n  \"StoryId\" : \"a58da017-05bb-48b0-c02f-08d80a57s458\",\n  \"StorySetId\" : \"d589e6df-4ac0-49c8-7dcd-08d80a548lfg\",\n  \"StoryName\" : \"Story about Love\",\n  \"StoryCoverImage\" : \"/heart.png\",\n  \"IsActive\" : true\n} ]";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<ArrayOfStories>(exampleJson)
                        : default(ArrayOfStories);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>Delete a story by ID</remarks>
        /// <param name="id">An ID to look up one or more values</param>
        /// <response code="200">OK</response>
        /// <response code="400">Bad Request</response>
        /// <response code="403">Access Denied</response>
        /// <response code="404">The resource is not found</response>
        [HttpDelete]
        [Route("/api/Story/{id}")]
        [ValidateModelState]
        [SwaggerOperation("ApiStoryIdDelete")]
        [SwaggerResponse(statusCode: 200, type: typeof(Story), description: "OK")]
        public virtual IActionResult ApiStoryIdDelete([FromRoute][Required]string id)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Story));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403);

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "{\n  \"StoryId\" : \"f89da017-05bb-48b0-c02f-08d80a57a67b\",\n  \"StorySetId\" : \"e520e6df-4ac0-49c8-7dcd-08d80a569eba\",\n  \"StoryName\" : \"Story about childhood\",\n  \"StoryCoverImage\" : \"/kids.png\",\n  \"IsActive\" : true\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<Story>(exampleJson)
                        : default(Story);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>Get a Story by ID</remarks>
        /// <param name="id">An ID to look up one or more values</param>
        /// <response code="200">OK</response>
        /// <response code="400">Bad Request</response>
        /// <response code="403">Access Denied</response>
        /// <response code="404">The resource is not found</response>
        [HttpGet]
        [Route("/api/Story/{id}")]
        [ValidateModelState]
        [SwaggerOperation("ApiStoryIdGet")]
        [SwaggerResponse(statusCode: 200, type: typeof(Story), description: "OK")]
        public virtual IActionResult ApiStoryIdGet([FromRoute][Required]string id)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Story));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403);

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "{\n  \"StoryId\" : \"f89da017-05bb-48b0-c02f-08d80a57a67b\",\n  \"StorySetId\" : \"e520e6df-4ac0-49c8-7dcd-08d80a569eba\",\n  \"StoryName\" : \"Story about childhood\",\n  \"StoryCoverImage\" : \"/kids.png\",\n  \"IsActive\" : true\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<Story>(exampleJson)
                        : default(Story);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>Add a new story</remarks>
        /// <param name="body"></param>
        /// <response code="200">OK</response>
        /// <response code="400">Bad Request</response>
        /// <response code="403">Access Denied</response>
        /// <response code="404">The resource is not found</response>
        [HttpPost]
        [Route("/api/Story")]
        [ValidateModelState]
        [SwaggerOperation("ApiStoryPost")]
        [SwaggerResponse(statusCode: 200, type: typeof(Story), description: "OK")]
        public virtual IActionResult ApiStoryPost([FromBody]Story body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Story));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403);

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "{\n  \"StoryId\" : \"f89da017-05bb-48b0-c02f-08d80a57a67b\",\n  \"StorySetId\" : \"e520e6df-4ac0-49c8-7dcd-08d80a569eba\",\n  \"StoryName\" : \"Story about childhood\",\n  \"StoryCoverImage\" : \"/kids.png\",\n  \"IsActive\" : true\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<Story>(exampleJson)
                        : default(Story);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>Edit existing stories</remarks>
        /// <param name="body"></param>
        /// <response code="200">OK</response>
        /// <response code="400">Bad Request</response>
        /// <response code="403">Access Denied</response>
        /// <response code="404">The resource is not found</response>
        [HttpPut]
        [Route("/api/Story")]
        [ValidateModelState]
        [SwaggerOperation("ApiStoryPut")]
        [SwaggerResponse(statusCode: 200, type: typeof(Story), description: "OK")]
        public virtual IActionResult ApiStoryPut([FromBody]Story body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Story));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403);

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "{\n  \"StoryId\" : \"f89da017-05bb-48b0-c02f-08d80a57a67b\",\n  \"StorySetId\" : \"e520e6df-4ac0-49c8-7dcd-08d80a569eba\",\n  \"StoryName\" : \"Story about childhood\",\n  \"StoryCoverImage\" : \"/kids.png\",\n  \"IsActive\" : true\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<Story>(exampleJson)
                        : default(Story);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}