# aspnetcore/src/IO.Swagger/Models/Story.cs
/*
 * Stories Collection
 *
 * Endpoints for updating Stories of Tales App
 *
 * OpenAPI spec version: 0.1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace IO.Swagger.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class Story : IEquatable<Story>
    { 
        /// <summary>
        /// ID of the Story
        /// </summary>
        /// <value>ID of the Story</value>
        [DataMember(Name="StoryId")]
        public string StoryId { get; set; }

        /// <summary>
        /// Gets or Sets StoryName
        /// </summary>
        [DataMember(Name="StoryName")]
        public string StoryName { get; set; }

        /// <summary>
        /// Story Preview image
        /// </summary>
        /// <value>Story Preview image</value>
        [DataMember(Name="StoryCoverImage")]
        public string StoryCoverImage { get; set; }

        /// <summary>
        /// Defines if the story is in use
        /// </summary>
        /// <value>Defines if the story is in use</value>
        [DataMember(Name="IsActive")]
        public bool? IsActive { get; set; }

        /// <summary>
        /// ID of the Story Set the story belongs to
        /// </summary>
        /// <value>ID of the Story Set the story belongs to</value>
        [DataMember(Name="StorySetId")]
        public string StorySetId { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class Story {\n");
            sb.Append("  StoryId: ").Append(StoryId).Append("\n");
            sb.Append("  StoryName: ").Append(StoryName).Append("\n");
            sb.Append("  StoryCoverImage: ").Append(StoryCoverImage).Append("\n");
            sb.Append("  IsActive: ").Append(IsActive).Append("\n");
            sb.Append("  StorySetId: ").Append(StorySetId).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((Story)obj);
        }

        /// <summary>
        /// Returns true if Story instances are equal
        /// </summary>
        /// <param name="other">Instance of Story to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(Story other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    StoryId == other.StoryId ||
                    StoryId != null &&
                    StoryId.Equals(other.StoryId)
                ) && 
                (
                    StoryName == other.StoryName ||
                    StoryName != null &&
                    StoryName.Equals(other.StoryName)
                ) && 
                (
                    StoryCoverImage == other.StoryCoverImage ||
                    StoryCoverImage != null &&
                    StoryCoverImage.Equals(other.StoryCoverImage)
                ) && 
                (
                    IsActive == other.IsActive ||
                    IsActive != null &&
                    IsActive.Equals(other.IsActive)
                ) && 
                (
                    StorySetId == other.StorySetId ||
                    StorySetId != null &&
                    StorySetId.Equals(other.StorySetId)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (StoryId != null)
                    hashCode = hashCode * 59 + StoryId.GetHashCode();
                    if (StoryName != null)
                    hashCode = hashCode * 59 + StoryName.GetHashCode();
                    if (StoryCoverImage != null)
                    hashCode = hashCode * 59 + StoryCoverImage.GetHashCode();
                    if (IsActive != null)
                    hashCode = hashCode * 59 + IsActive.GetHashCode();
                    if (StorySetId != null)
                    hashCode = hashCode * 59 + StorySetId.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(Story left, Story right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(Story left, Story right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}