# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Stories Collection",
    "description" : "Endpoints for updating Stories of Tales App",
    "version" : "0.1"
  },
  "servers" : [ {
    "url" : "https://localhost:44382"
  } ],
  "paths" : {
    "/api/Story" : {
      "put" : {
        "description" : "Edit existing stories",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Story"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Story"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "403" : {
            "description" : "Access Denied"
          },
          "404" : {
            "description" : "The resource is not found"
          }
        }
      },
      "post" : {
        "description" : "Add a new story",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Story"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Story"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "403" : {
            "description" : "Access Denied"
          },
          "404" : {
            "description" : "The resource is not found"
          }
        }
      }
    },
    "/api/Story/GetAllActiveStoriesOfSet/{id}" : {
      "get" : {
        "description" : "Get all active stories of a set by ID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "An ID to look up one or more values",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArrayOfStories"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "403" : {
            "description" : "Access Denied"
          },
          "404" : {
            "description" : "The resource is not found"
          }
        }
      }
    },
    "/api/Story/{id}" : {
      "get" : {
        "description" : "Get a Story by ID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "An ID to look up one or more values",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Story"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "403" : {
            "description" : "Access Denied"
          },
          "404" : {
            "description" : "The resource is not found"
          }
        }
      },
      "delete" : {
        "description" : "Delete a story by ID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "An ID to look up one or more values",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Story"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "403" : {
            "description" : "Access Denied"
          },
          "404" : {
            "description" : "The resource is not found"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "ArrayOfStories" : {
        "title" : "Stories array",
        "type" : "array",
        "example" : [ {
          "StoryId" : "f89da017-05bb-48b0-c02f-08d80a57a67b",
          "StorySetId" : "e520e6df-4ac0-49c8-7dcd-08d80a569eba",
          "StoryName" : "Story about childhood",
          "StoryCoverImage" : "/kids.png",
          "IsActive" : true
        }, {
          "StoryId" : "a58da017-05bb-48b0-c02f-08d80a57s458",
          "StorySetId" : "d589e6df-4ac0-49c8-7dcd-08d80a548lfg",
          "StoryName" : "Story about Love",
          "StoryCoverImage" : "/heart.png",
          "IsActive" : true
        } ],
        "items" : {
          "$ref" : "#/components/schemas/ArrayOfStories_inner"
        }
      },
      "Story" : {
        "title" : "Story",
        "required" : [ "StoryId StoryName StoryCoverImage IsActive StorySetId" ],
        "type" : "object",
        "properties" : {
          "StoryId" : {
            "type" : "string",
            "description" : "ID of the Story"
          },
          "StoryName" : {
            "type" : "string"
          },
          "StoryCoverImage" : {
            "type" : "string",
            "description" : "Story Preview image"
          },
          "IsActive" : {
            "type" : "boolean",
            "description" : "Defines if the story is in use"
          },
          "StorySetId" : {
            "type" : "string",
            "description" : "ID of the Story Set the story belongs to"
          }
        },
        "example" : {
          "StoryId" : "f89da017-05bb-48b0-c02f-08d80a57a67b",
          "StorySetId" : "e520e6df-4ac0-49c8-7dcd-08d80a569eba",
          "StoryName" : "Story about childhood",
          "StoryCoverImage" : "/kids.png",
          "IsActive" : true
        }
      },
      "ArrayOfStories_inner" : {
        "type" : "object",
        "properties" : {
          "StoryId" : {
            "type" : "string",
            "description" : "ID of the Story"
          },
          "StoryName" : {
            "type" : "string"
          },
          "StoryCoverImage" : {
            "type" : "string",
            "description" : "Story Preview image"
          },
          "IsActive" : {
            "type" : "boolean",
            "description" : "Defines if the story is in use"
          },
          "StorySetId" : {
            "type" : "string",
            "description" : "ID of the Story Set the story belongs to"
          }
        }
      }
    },
    "responses" : {
      "BadRequest" : {
        "description" : "Bad Request"
      },
      "AccessDenied" : {
        "description" : "Access Denied"
      },
      "NotFound" : {
        "description" : "The resource is not found"
      }
    },
    "parameters" : {
      "lookUpById" : {
        "name" : "id",
        "in" : "path",
        "description" : "An ID to look up one or more values",
        "required" : true,
        "style" : "simple",
        "explode" : false,
        "schema" : {
          "type" : "string"
        }
      }
    }
  }
}